+++ src/number-field/number-field.js
@@ -96,6 +96,7 @@ export class RowanNumberField extends BaseElement {
   #inputId = "";
   #autoInvalid = false;
   #paintInvalid = false;
+  #ignoreInvalidEvent = false;
 
   connectedCallback() {
     super.connectedCallback();
@@ -231,16 +232,22 @@ export class RowanNumberField extends BaseElement {
   }
 
   checkValidity() {
-    if (this.internals && typeof this.internals.checkValidity === "function") {
-      return this.internals.checkValidity();
-    }
+    this.#ignoreInvalidEvent = true;
+    try {
+      if (this.internals && typeof this.internals.checkValidity === "function") {
+        return this.internals.checkValidity();
+      }
 
-    return this.#input ? this.#input.checkValidity() : true;
+      return this.#input ? this.#input.checkValidity() : true;
+    } finally {
+      this.#ignoreInvalidEvent = false;
+    }
   }
 
   reportValidity() {
     this.#paintInvalid = true;
     this.#syncValidity();
+    this.#paintHostIfInvalid();
     this.#applyDefaultA11y();
 
     if (this.internals && typeof this.internals.reportValidity === "function") {
@@ -304,6 +311,8 @@ export class RowanNumberField extends BaseElement {
         this.#paintInvalid = true;
         this.requestRender();
       });
+
+      this.listen(this, "invalid", () => this.#onInvalid());
 
       this.listen(this.#decrementButton, "click", () => {
         this.#nudge(-1);
@@ -395,6 +404,20 @@ export class RowanNumberField extends BaseElement {
     emit(this, "rowan-change", {
       value: this.value,
     });
+  }
+
+  #onInvalid() {
+    if (this.#ignoreInvalidEvent) return;
+    this.#paintInvalid = true;
+    this.#syncValidity();
+    this.#paintHostIfInvalid();
+    this.#applyDefaultA11y();
+  }
+
+  #paintHostIfInvalid() {
+    if (this.internals && this.internals.validity && !this.internals.validity.valid) {
+      this.#setAutoInvalid(true);
+    }
   }
 
   #syncFormValue() {