+++ src/textarea/textarea.js
@@ -51,6 +51,7 @@ export class RowanTextarea extends BaseElement {
   #inputId = "";
   #autoInvalid = false;
   #paintInvalid = false;
+  #ignoreInvalidEvent = false;
 
   connectedCallback() {
     super.connectedCallback();
@@ -163,16 +164,22 @@ export class RowanTextarea extends BaseElement {
   }
 
   checkValidity() {
-    if (this.internals && typeof this.internals.checkValidity === "function") {
-      return this.internals.checkValidity();
-    }
+    this.#ignoreInvalidEvent = true;
+    try {
+      if (this.internals && typeof this.internals.checkValidity === "function") {
+        return this.internals.checkValidity();
+      }
 
-    return this.#input ? this.#input.checkValidity() : true;
+      return this.#input ? this.#input.checkValidity() : true;
+    } finally {
+      this.#ignoreInvalidEvent = false;
+    }
   }
 
   reportValidity() {
     this.#paintInvalid = true;
     this.#syncValidity();
+    this.#paintHostIfInvalid();
     this.#applyDefaultA11y();
 
     if (this.internals && typeof this.internals.reportValidity === "function") {
@@ -211,6 +218,8 @@ export class RowanTextarea extends BaseElement {
         this.#syncValidity();
         this.#applyDefaultA11y();
       });
+
+      this.listen(this, "invalid", () => this.#onInvalid());
     }
 
     this.#input.id = this.#inputId;
@@ -235,6 +244,20 @@ export class RowanTextarea extends BaseElement {
     this.#syncFormValue();
     this.#syncValidity();
     this.#applyDefaultA11y();
+  }
+
+  #onInvalid() {
+    if (this.#ignoreInvalidEvent) return;
+    this.#paintInvalid = true;
+    this.#syncValidity();
+    this.#paintHostIfInvalid();
+    this.#applyDefaultA11y();
+  }
+
+  #paintHostIfInvalid() {
+    if (this.internals && this.internals.validity && !this.internals.validity.valid) {
+      this.#setAutoInvalid(true);
+    }
   }
 
   #syncFormValue() {